WIP state refactor.
- Continue working from test_router.py -> test_nimbus_solve
