+++ desdeo/api/routers/utopia.py
@@ -1,4 +1,5 @@
 """Utopia router"""
+
 import json
 from typing import Annotated
 
@@ -6,11 +7,11 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import Session, select
 
 from desdeo.api.db import get_session
-from desdeo.api.models import ForestProblemMetaData, ProblemMetaDataDB, User, UtopiaRequest, UtopiaResponse
-from desdeo.api.models.state import StateDB
+from desdeo.api.models import ForestProblemMetaData, ProblemMetaDataDB, User, UtopiaRequest, UtopiaResponse, StateDB
 from desdeo.api.routers.user_authentication import get_current_user
 
 router = APIRouter(prefix="/utopia")
+
 
 @router.post("/")
 def get_utopia_data(
@@ -19,53 +20,52 @@ def get_utopia_data(
     session: Annotated[Session, Depends(get_session)],
 ) -> UtopiaResponse:
     """Request and receive the Utopia map corresponding to the decision variables sent. Can be just the optimal_variables form a SolverResult
-    
+
     Args:
         request (UtopiaRequest): the set of decision variables and problem for which the utopia forest map is requested for.
         user (Annotated[User, Depend(get_current_user)]) the current user
         session (Annotated[Session, Depends(get_session)]) the current database session
 
     Raises:
-        HTTPException: 
+        HTTPException:
 
     Returns:
         UtopiaResponse: the map for the forest, to be rendered in frontend
     """
 
-    empty_response = UtopiaResponse(
-            is_utopia=False,
-            map_name="",
-            map_json={},
-            options={},
-            description="",
-            years=[]
-        )
+    empty_response = UtopiaResponse(is_utopia=False, map_name="", map_json={}, options={}, description="", years=[])
 
     state = session.exec(select(StateDB).where(StateDB.id == request.solution.address_state)).first()
-    if state is None or not hasattr(state, 'state'):
+    if state is None or not hasattr(state, "state"):
         return empty_response
 
     # Check if solver_results exists and has the needed index
-    if (not hasattr(state.state, 'solver_results') or
-            request.solution.address_result >= len(state.state.solver_results) or
-            state.state.solver_results[request.solution.address_result] is None):
+    if (
+        not hasattr(state.state, "solver_results")
+        or request.solution.address_result >= len(state.state.solver_results)
+        or state.state.solver_results[request.solution.address_result] is None
+    ):
         return empty_response
 
     result = state.state.solver_results[request.solution.address_result]
-    if not hasattr(result, 'optimal_variables') or not result.optimal_variables:
+    if not hasattr(result, "optimal_variables") or not result.optimal_variables:
         return empty_response
 
-    decision_variables = result.optimal_variables #expects a list of variables, won't work without.
+    decision_variables = result.optimal_variables  # expects a list of variables, won't work without.
 
-    from_db_metadata = session.exec(select(ProblemMetaDataDB).where(ProblemMetaDataDB.problem_id == request.problem_id)).first()
+    from_db_metadata = session.exec(
+        select(ProblemMetaDataDB).where(ProblemMetaDataDB.problem_id == request.problem_id)
+    ).first()
     if from_db_metadata is None:
         return empty_response
-    
+
     # Get the last instance of forest related metadata from the database. If for some reason there's more than one forest metadata, return the latest.
-    forest_metadata: ForestProblemMetaData = [metadata for metadata in from_db_metadata.data if metadata.metadata_type == "forest_problem_metadata"][-1]
+    forest_metadata: ForestProblemMetaData = [
+        metadata for metadata in from_db_metadata.data if metadata.metadata_type == "forest_problem_metadata"
+    ][-1]
     if forest_metadata is None:
         return empty_response
-    
+
     # Figure out the treatments from the decision variables and utopia data
 
     def treatment_index(part: str) -> str:
@@ -197,10 +197,10 @@ def get_utopia_data(
 
     # Let's also generate a nice description for the map
     map_description = (
-        f"Income from harvesting in the first period {int(decision_variables["P_1"])}€.\n" # Will complain about P_X if there is no P_X.
-        + f"Income from harvesting in the second period {int(decision_variables["P_2"])}€.\n" # See the comment about what is expected
-        + f"Income from harvesting in the third period {int(decision_variables["P_3"])}€.\n" # from the UtopiaRequest.
-        + f"The discounted value of the remaining forest at the end of the plan {int(decision_variables["V_end"])}€."
+        f"Income from harvesting in the first period {int(decision_variables['P_1'])}€.\n"  # Will complain about P_X if there is no P_X.
+        + f"Income from harvesting in the second period {int(decision_variables['P_2'])}€.\n"  # See the comment about what is expected
+        + f"Income from harvesting in the third period {int(decision_variables['P_3'])}€.\n"  # from the UtopiaRequest.
+        + f"The discounted value of the remaining forest at the end of the plan {int(decision_variables['V_end'])}€."
     )
 
     return UtopiaResponse(
@@ -211,4 +211,3 @@ def get_utopia_data(
         description=map_description,
         years=forest_metadata.years,
     )
-
